+++ src/Console/CommandException.php
@@ -2,14 +2,16 @@
 
 namespace Solital\Core\Console;
 
+use Throwable;
+
 class CommandException extends \Exception
 {
     /**
      * @param string $message
      * @param int $code
-     * @param \Throwable|null $previous
+     * @param Throwable|null $previous
      */
-    public function __construct(string $message = "", $code = 0, \Throwable $previous = null)
+    public function __construct(string $message = "", $code = 0, ?Throwable $previous = null)
     {
         parent::__construct($message, $code, $previous);
     }