+++ src/Console/tests/Commands/Custom.php
@@ -5,7 +5,7 @@ namespace Solital\Core\Console\Tests\Commands;
 use Solital\Core\Console\Command;
 use Solital\Core\Console\Interface\CommandInterface;
 
-class CustomTest extends Command implements CommandInterface
+class Custom extends Command implements CommandInterface
 {
     protected string $command = "create:testsecond";
     protected array $arguments = ["name"];