+++ src/Console/tests/VinciTest.php
@@ -4,42 +4,49 @@ namespace Solital\Core\Console\Tests;
 
 use PHPUnit\Framework\TestCase;
 use Solital\Core\Console\{Command, CommandException};
-use Solital\Core\Console\Tests\ExtendCommands\{ExtendedCommandsTest, OtherExtendComandTest};
+use Solital\Core\Console\Tests\ExtendCommands\{ExtendedCommands, OtherExtendComand};
 
 class VinciTest extends TestCase
 {
+    private Command $command;
+
     private array $class_commands = [
-        ExtendedCommandsTest::class,
-        OtherExtendComandTest::class
+        ExtendedCommands::class,
+        OtherExtendComand::class
     ];
+
+    public function setUp() : void
+    {
+        $this->command = new Command($this->class_commands);
+    }
 
     public function testReadWithoutArgument()
     {
-        $res = (new Command($this->class_commands))->read('user:test');
+        $res = $this->command->read('user:test');
         $this->assertEquals("user:test - Command OK\r\n", $res);
     }
 
     public function testReadWithOptionEmpty()
     {
-        $res = (new Command($this->class_commands))->read('create:test', ['', '', '--option']);
+        $res = $this->command->read('create:test', ['', '', '--option']);
         $this->assertEquals(true, $res);
     }
 
     public function testReadWithOptionValue()
     {
-        $res = (new Command($this->class_commands))->read('create:test', ['', '', '--witharg=accept']);
+        $res = $this->command->read('create:test', ['', '', '--witharg=accept']);
         $this->assertEquals("accept", $res);
     }
 
     public function testReadArgument()
     {
-        $res = (new Command($this->class_commands))->read('create:testsecond', ['', '', 'Vinci']);
+        $res = $this->command->read('create:testsecond', ['', '', 'Vinci']);
         $this->assertEquals("Vinci", $res);
     }
 
     public function testExceptionSameCommand()
     {
         $this->expectException(CommandException::class);
-        (new Command($this->class_commands))->read('same:command');
+        $this->command->read('same:command');
     }
 }